requirements.txt
keboola.component>=1.6.0
anthropic>=0.40.0
pandas>=2.0.0
pydantic>=2.0.0
